Simplify FileSystem.Create, update Bazel deps.
To create an object in the file system, implementers would have to also
implement a customer Writer interface, with special behaviour around a Sync()
function. This is prone to errors, and is a dirty hack for keeping consistency.
The Writer interface has been completely removed, and the Create function now
requires a Reader.

Also update Bazel dependencies.